requirements.txt
einops
roma
accelerate
smplx
chumpy
decord==0.6.0
diffusers==0.32.0
dna==0.0.1
gradio==4.43.0
gfpgan==1.3.8
gsplat==1.4.0
huggingface_hub==0.23.2
imageio==2.34.1
imageio-ffmpeg
jaxtyping==0.2.38
kiui==0.2.14
kornia==0.7.2
loguru==0.7.3
lpips==0.1.4
matplotlib==3.5.3
megfile==4.1.0.post2
numpy==1.23.0
omegaconf==2.3.0
open3d==0.19.0
opencv_python
opencv_python_headless
Pillow==10.4.0
pygltflib==1.16.2
pyrender==0.1.45
PyYAML==6.0.1
rembg==2.0.63
Requests==2.32.3
scipy
spaces
setuptools==74.0.0
taming_transformers_rom1504==0.0.6
timm==1.0.15
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
trimesh==4.4.9
typeguard==2.13.3
xatlas==0.0.9
xformers==0.0.26.post1

onnx2torch
face-alignment==1.4.1
json-tricks
av==12.1.0